Decouple notifications logic from mail logic.
